requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
sentence-transformers>=2.2.0
